# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HaulLedger API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  expense: "expense",
  income: "income",
  fuel: "fuel",
  trip: "trip",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  amount: number;
  date: string;
}

export interface Summary {
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  weeklyMiles: number;
  mileageGoal: number;
  recentActivity: ActivityItem[];
}

export interface Expense {
  id: number;
  date: string;
  merchant: string;
  category: string;
  amount: number;
  notes?: string | null;
  receiptUrl?: string | null;
  truckId?: number | null;
  gallons?: number | null;
  pricePerGallon?: number | null;
  jurisdiction?: string | null;
  createdAt: string;
}

export interface CreateExpenseInput {
  date: string;
  merchant: string;
  category: string;
  amount: number;
  notes?: string | null;
  receiptUrl?: string | null;
  truckId?: number | null;
  gallons?: number | null;
  pricePerGallon?: number | null;
  jurisdiction?: string | null;
}

export interface Income {
  id: number;
  date: string;
  source: string;
  amount: number;
  trailerNumber?: string | null;
  routeName?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateIncomeInput {
  date: string;
  source: string;
  amount: number;
  trailerNumber?: string | null;
  routeName?: string | null;
  notes?: string | null;
}

export interface FuelEntry {
  id: number;
  date: string;
  vendor: string;
  gallons: number;
  pricePerGallon: number;
  jurisdiction: string;
  totalAmount: number;
  truckId?: number | null;
  createdAt: string;
}

export interface CreateFuelEntryInput {
  date: string;
  vendor: string;
  gallons: number;
  pricePerGallon: number;
  jurisdiction: string;
  totalAmount: number;
  truckId?: number | null;
}

export interface Trip {
  id: number;
  date: string;
  startOdometer: number;
  endOdometer: number;
  loadedMiles: number;
  emptyMiles: number;
  jurisdiction: string;
  truckId?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateTripInput {
  date: string;
  startOdometer: number;
  endOdometer: number;
  loadedMiles: number;
  emptyMiles: number;
  jurisdiction: string;
  truckId?: number | null;
  notes?: string | null;
}

export type AssetType = (typeof AssetType)[keyof typeof AssetType];

export const AssetType = {
  Truck: "Truck",
  Trailer: "Trailer",
} as const;

export interface Asset {
  id: number;
  type: AssetType;
  vin: string;
  plate: string;
  year: number;
  make: string;
  model: string;
  totalMiles: number;
  createdAt: string;
}

export type CreateAssetInputType =
  (typeof CreateAssetInputType)[keyof typeof CreateAssetInputType];

export const CreateAssetInputType = {
  Truck: "Truck",
  Trailer: "Trailer",
} as const;

export interface CreateAssetInput {
  type: CreateAssetInputType;
  vin: string;
  plate: string;
  year: number;
  make: string;
  model: string;
}

export interface SavedRoute {
  id: number;
  name: string;
  origin: string;
  destination: string;
  standardRate: number;
  createdAt: string;
}

export interface CreateSavedRouteInput {
  name: string;
  origin: string;
  destination: string;
  standardRate: number;
}

export interface QuickExpense {
  id: number;
  label: string;
  category: string;
  defaultAmount: number;
  createdAt: string;
}

export interface CreateQuickExpenseInput {
  label: string;
  category: string;
  defaultAmount: number;
}

export interface ParsedReceipt {
  merchant?: string | null;
  date?: string | null;
  amount?: number | null;
  gallons?: number | null;
  pricePerGallon?: number | null;
  jurisdiction?: string | null;
  category?: string | null;
}

export type GetExpensesParams = {
  category?: string;
  week?: boolean;
  search?: string;
};

export type GetIncomeParams = {
  week?: boolean;
};

export type ProcessReceiptBody = {
  imageBase64: string;
};