# lib/api-zod/src/generated/types/createAssetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HaulLedger API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAssetInputType } from "./createAssetInputType";

export interface CreateAssetInput {
  type: CreateAssetInputType;
  vin: string;
  plate: string;
  year: number;
  make: string;
  model: string;
}