# lib/api-zod/src/generated/types/createExpenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HaulLedger API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseInput {
  date: string;
  merchant: string;
  category: string;
  amount: number;
  notes?: string | null;
  receiptUrl?: string | null;
  truckId?: number | null;
  gallons?: number | null;
  pricePerGallon?: number | null;
  jurisdiction?: string | null;
}