# lib/api-zod/src/generated/types/createIncomeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HaulLedger API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIncomeInput {
  date: string;
  source: string;
  amount: number;
  trailerNumber?: string | null;
  routeName?: string | null;
  notes?: string | null;
}