# lib/api-zod/src/generated/types/createSavedRouteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HaulLedger API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSavedRouteInput {
  name: string;
  origin: string;
  destination: string;
  standardRate: number;
}