# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HaulLedger API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Trip {
  id: number;
  date: string;
  startOdometer: number;
  endOdometer: number;
  loadedMiles: number;
  emptyMiles: number;
  jurisdiction: string;
  truckId?: number | null;
  notes?: string | null;
  createdAt: string;
}